move location type to tour package
